Refactor metadata and archive old results
Restructure the run metadata to distinguish the code used from
the compiler/interpreter backend and add fields for versions as well.

N.B. --backend now refers to C++/Julia/Python, not to the code.
used, which is --code. And the options there are updated to be
the packages used for the code, i.e., JetReconstruction Fastjet
AkTPython AkTNumPy.

As the metadata format is now so different from before, and not
backwards compatible, the old results directory is moved to
results-v0.